requirements.txt
amqp==5.3.1
anyio==4.11.0
asgiref==3.9.2
attrs==25.3.0
billiard==4.2.2
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
cryptography==46.0.1
Django==5.2.6
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-filter==25.1
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
et_xmlfile==2.0.0
h11==0.16.0
hiredis==3.2.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
Markdown==3.9
nanoid==2.0.0
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
pylibmc==1.6.3
pymemcache==4.0.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
